// client/src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      events: {
        Row: {
          attendance_points: number | null
          created_at: string | null
          created_by: string | null
          current_capacity: number | null
          description: string | null
          end_time: string | null
          id: string
          image_url: string | null
          location: string | null
          loyalty_points_price: number | null
          max_capacity: number | null
          price_cents: number | null
          qr_code_generated_at: string | null
          qr_code_generated_by: string | null
          qr_code_token: string | null
          start_time: string
          status: Database["public"]["Enums"]["event_status"] | null
          title: string
          updated_at: string | null
        }
        Insert: {
          attendance_points?: number | null
          created_at?: string | null
          created_by?: string | null
          current_capacity?: number | null
          description?: string | null
          end_time?: string | null
          id?: string
          image_url?: string | null
          location?: string | null
          loyalty_points_price?: number | null
          max_capacity?: number | null
          price_cents?: number | null
          qr_code_generated_at?: string | null
          qr_code_generated_by?: string | null
          qr_code_token?: string | null
          start_time: string
          status?: Database["public"]["Enums"]["event_status"] | null
          title: string
          updated_at?: string | null
        }
        Update: {
          attendance_points?: number | null
          created_at?: string | null
          created_by?: string | null
          current_capacity?: number | null
          description?: string | null
          end_time?: string | null
          id?: string
          image_url?: string | null
          location?: string | null
          loyalty_points_price?: number | null
          max_capacity?: number | null
          price_cents?: number | null
          qr_code_generated_at?: string | null
          qr_code_generated_by?: string | null
          qr_code_token?: string | null
          start_time?: string
          status?: Database["public"]["Enums"]["event_status"] | null
          title?: string
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
      profiles: {
        Row: {
          available_loyalty_points: number | null
          avatar_url: string | null
          bio: string | null
          created_at: string | null
          full_name: string | null
          id: string
          instagram_handle: string | null
          is_active: boolean
          location: string | null
          public_key: string | null
          total_loyalty_points: number | null
          updated_at: string | null
          username: string | null
        }
        Insert: {
          available_loyalty_points?: number | null
          avatar_url?: string | null
          bio?: string | null
          created_at?: string | null
          full_name?: string | null
          id: string
          instagram_handle?: string | null
          is_active?: boolean
          location?: string | null
          public_key?: string | null
          total_loyalty_points?: number | null
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          available_loyalty_points?: number | null
          avatar_url?: string | null
          bio?: string | null
          created_at?: string | null
          full_name?: string | null
          id?: string
          instagram_handle?: string | null
          is_active?: boolean
          location?: string | null
          public_key?: string | null
          total_loyalty_points?: number | null
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never